web_ui/requirements.txt
flask==2.3.3
python-dotenv==1.0.0
keepa==1.3.16
requests==2.31.0
